### 2025-03-21 #7933 @editorialbot
[PRE REVIEW]: AI-ANNE: (A) (N)eural (N)et for (E)xploration
**Submitting author:** @statistical-thinking  (<a href="http://orcid.org/0000-0003-1011-5517">Prof. Dr. habil. Dennis Klinkhammer</a>)
**Repository:** https://github.com/statistical-thinking/KI.ENNA
**Branch with paper.md** (empty if default branch):
**Version:** 2.0

Thanks for submitting your paper to JOSS @statistical-thinking . **Currently, there isn't a JOSS editor assigned** to your paper.
