+++ paper/paper.bib
@@ -37,7 +37,8 @@
   month     =  jan,
   year      =  2020,
   copyright = "https://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = {10.3390/su12020492}
 }
 
 @ARTICLE{Collier2024-ld,
@@ -50,7 +51,8 @@
   volume    =  35,
   number    =  3,
   pages     = "325--376",
-  year      =  2024
+  year      =  2024,
+  doi       = {10.62273/spyc4248}
 }
 
 @ARTICLE{Delnevo2023-ux,
@@ -66,7 +68,8 @@
   month     =  jul,
   year      =  2023,
   copyright = "http://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = {10.1016/j.iot.2023.100729}
 }
 
 @ARTICLE{Haque2023-du,
@@ -82,7 +85,8 @@
   month     =  jan,
   year      =  2023,
   copyright = "http://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = {10.1016/j.techfore.2022.122120}
 }
 
 @ARTICLE{Kong2022-sk,
@@ -98,7 +102,8 @@
   month     =  aug,
   year      =  2022,
   copyright = "http://creativecommons.org/licenses/by-nc-nd/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = {10.1016/j.chbr.2022.100223}
 }
 
 @ARTICLE{LeCun2015-hn,
@@ -125,7 +130,8 @@
   pages     = "436--444",
   month     =  may,
   year      =  2015,
-  language  = "en"
+  language  = "en",
+  doi       = {10.1038/nature14539}
 }
 
 @ARTICLE{Li2021-bn,
@@ -140,7 +146,8 @@
   month     =  jul,
   year      =  2021,
   copyright = "http://creativecommons.org/licenses/by-nc-nd/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = {10.1016/j.jmsy.2021.07.007}
 }
 
 @ARTICLE{Meske2022-ke,
@@ -155,7 +162,8 @@
   pages     = "53--63",
   month     =  jan,
   year      =  2022,
-  language  = "en"
+  language  = "en",
+  doi       = {10.1080/10580530.2020.1849465}
 }
 
 @ARTICLE{Ray2022-oq,
@@ -169,7 +177,8 @@
   month     =  apr,
   year      =  2022,
   copyright = "http://creativecommons.org/licenses/by-nc-nd/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = {10.1016/j.jksuci.2021.11.019}
 }
 
 @ARTICLE{Sakr2021-mc,
@@ -205,7 +214,8 @@
   month     =  oct,
   year      =  2021,
   copyright = "https://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = {10.3390/en14206636}
 }
 
 @ARTICLE{Scherer2016-fy,
@@ -220,7 +230,8 @@
   year      =  2016,
   keywords  = "computational thinking; computer programming; creativity;
                problem solving; transfer of learning",
-  language  = "en"
+  language  = "en",
+  doi       = {10.3389/fpsyg.2016.01390}
 }
 
 @ARTICLE{Schmidt2020-ak,
@@ -233,7 +244,8 @@
   pages     = "260--278",
   month     =  oct,
   year      =  2020,
-  language  = "en"
+  language  = "en",
+  doi       = {10.1080/12460125.2020.1819094}
[…]
