### 2025-03-21 #7933 @statistical-thinking
Title is supposed to be "AI-ANNE: (A) (N)eural (N)et for (E)xploration" and not "Explainable Artificial Intelligence with MicroPython: Lightweight Neural Networks for Students’ Deeper Learning"...

Previous issues have been fixed.
