+++ paper/paper.md
@@ -18,11 +18,11 @@ bibliography: paper.bib
 
 # Summary
 
-Machine learning and deep learning are increasingly driving innovation across various fields (LeCun et al. 2015) and can also be transferred on microcontrollers (Ray 2022), for example to process sensor data (Cioffi et al. 2020). Since microcontrollers have limited computational resources (Delnevo et al. 2023), training or developing neural networks on microcontrollers remains a complex challenge (Wulfert et al. 2024). However, MicroPython as programming language enables a resource-efficient implementation of neural networks on microcontrollers and provides insights into the fundamentals of neural networks in terms of explainable artificial intelligence (Hague et al. 2023; Meske et al. 2022), while coding from scratch in MicroPython also allows for a streamlined and tailored approach (Delnevo et al. 2023). AI-ANNE: (A) (N)eural (N)et for (E)xploration provides a framework that transfers neural networks from Python to MicroPython and enables the application of pre-trained TensorFlow and Keras models on microcontrollers. Furthermore, it enables users to explore the performance of neural networks while simultaneously the number of neurons and layers as well as the underlying activation functions can be adjusted easily in MicroPython, which makes it also suitable for didactic application (Collier & Powell 2024; Meske et al. 2022; Verma et al. 2022; Scherer 2016).
+Machine learning and deep learning are increasingly driving innovation across various fields [@LeCun2015-hn] and can also be transferred on microcontrollers [@Ray2022-oq], for example to process sensor data [@Cioffi2020-lq]. Since microcontrollers have limited computational resources [@Delnevo2023-ux], training or developing neural networks on microcontrollers remains a complex challenge [@Wulfert2024-qf]. However, MicroPython as programming language enables a resource-efficient implementation of neural networks on microcontrollers and provides insights into the fundamentals of neural networks in terms of explainable artificial intelligence [@Haque2023-du; @Meske2022-ke], while coding from scratch in MicroPython also allows for a streamlined and tailored approach [@Delnevo2023-ux]. AI-ANNE: (A) (N)eural (N)et for (E)xploration provides a framework that transfers neural networks from Python to MicroPython and enables the application of pre-trained TensorFlow and Keras models on microcontrollers. Furthermore, it enables users to explore the performance of neural networks while simultaneously the number of neurons and layers as well as the underlying activation functions can be adjusted easily in MicroPython, which makes it also suitable for didactic application [@Collier2024-ld; @Meske2022-ke; @Verma2022-ie; @Scherer2016-fy].
 
 # Statement of need
 
[…]
