### 2025-03-22 #7933 @statistical-thinking
**Suggestions for potential reviewers** that have knowledge about embedded systems and could be familiar with MicroPython and microcontrollers in Order to check ai-anne-b.py (works even without a Raspberry Pi Pico):

1) samiralavi
2) SamMachariaPhD
3) ixjlyons
