+++ README.md
@@ -17,8 +17,8 @@ An ongoing research project is testing the use of AI-ENNA to detect hazards in t
 The “python” folder contains a data set in CSV format as well as Jupyter notebooks for two neural networks for classification.
 The data sets were each used as the basis for training the neural networks with TensorFlow in Python.
 The parameters of the neural networks were then extracted as a TXT file.
-The “micropython” folder contains the code for initializing the microcontroller.
-This is copied to the microcontroller via Thonny and the microcontroller is then restarted.
+The “micropython” folder contains the **code for initializing the microcontroller**.
+This is **copied to the microcontroller** via Thonny and the microcontroller is then restarted.
 The IRIS data set and the DIABETES data set are available as example data sets.
 'ReLU', 'Leaky ReLU', 'Sigmoid', 'Softmax' and 'Tanh' are available as activation functions.
 If required, other classification data can be used for training in Python.
@@ -26,9 +26,9 @@ For other data sets, the parameters of the neural networks must be adjusted acco
 AI-ANNE-B(asic) works directly in Thonny and does not require a display or RGB matrix.
 
 # License & Acknowledgment
-The current version 2.0 of AI-ANNE is available under MIT License.
+The current version 2.0 of AI-ANNE is available under **MIT License**.
 Version 1.0 from 2021 was available under the German Free Software License.
-Special thanks @tonygod2 from instructables.com for his tutorials on getting the display up and running.
+Special thanks **@tonygod2** from instructables.com for his tutorials on getting the display up and running.
 
 ------------------------------------------------------------------------------------------
 
@@ -51,8 +51,8 @@ In einem laufenden Forschungsvorhaben wird der Einsatz von KI-ENNA zur Detektion
 Im Ordner "python" befindet sich jeweils ein Datensatz im CSV-Format sowie Jupyter Notebooks für zwei Neuronale Netze zur Klassifikation.
 Die Datensätze wurden jeweils als Grundlage für das Training der Neuronalen Netze mit TensorFlow in Python verwendet.
 Die Parameter der Neuronalen Netze wurden daraufhin als TXT-Datei extrahiert.
-Im Ordner "micropython" befindet sich der Code zur Initialisierung des Microcontrollers.
-Dieser wird via Thonny auf den Microcontroller kopiert und dieser anschließend neu gestartet.
+Im Ordner "micropython" befindet sich der **Code zur Initialisierung des Microcontrollers**.
+Dieser wird via Thonny **auf den Microcontroller kopiert** und dieser anschließend neu gestartet.
 Als Beispieldatensätze stehen der IRIS Datensatz und der DIABETES Datensatz zur Verfügung.
 'ReLU', 'Leaky ReLU', 'Sigmoid', 'Softmax' und 'Tanh' stehen als Aktivierungsfunktionen zur Verfügung.
 Bei Bedarf können andere Klassifikationsdaten für das Training in Python verwendet werden.
@@ -60,9 +60,9 @@ Bei anderen Datensätzen sind die Parameter der Neuronalen Netze entsprechend an
 KI-ENNA-B(asic) funktioniert direkt in Thonny und benötigt weder Display noch RGB-Matrix.
 
 # Lizenz & Danksagung
-In der aktuellen Version 2.0 steht KI-ENNA unter der MIT Lizens zur Verfügung.
+In der aktuellen Version 2.0 steht KI-ENNA unter der **MIT Lizens** zur Verfügung.
 Version 1.0 von 2021 war noch unter der German Free Software License verfgügbar.
-Besonderer Dank gilt @tonygod2 von instructables.com für seine hilfreichen Display-Tutorials.
+Besonderer Dank gilt **@tonygod2** von instructables.com für seine hilfreichen Display-Tutorials.
 
 ------------------------------------------------------------------------------------------
 
