### 2025-04-11 #8039 @editorialbot
[REVIEW]: AI-ANNE: (A) (N)eural (N)et for (E)xploration
**Submitting author:** @statistical-thinking  (<a href="http://orcid.org/0000-0003-1011-5517">Prof. Dr. habil. Dennis Klinkhammer</a>)
**Repository:** https://github.com/statistical-thinking/KI.ENNA
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.1
