### 2025-04-14 #8039 @kalpan80
@osorensen - Noted a typo in the paper, Page 2, paragraph 2, Line 3. Lightweight is spelled incorrectly as leightweight.
