### 2025-04-15 #8039 @statistical-thinking
@kalpan80 This made me sweat: Couldn't find the word “leightweight” anywhere in the JOSS paper... then I found it in the ArXiv PrePrint / Tutorial. THX :-)
