+++ README.md
@@ -17,13 +17,13 @@ An ongoing research project is testing the use of AI-ENNA to detect hazards in t
 The “python” folder contains a data set in CSV format as well as Jupyter notebooks for two neural networks for classification.
 The data sets were each used as the basis for training the neural networks with TensorFlow in Python.
 The parameters of the neural networks were then extracted as a TXT file.
-The “micropython” folder contains the **code for initializing the microcontroller**.
+The “micropython” folder contains the **code for initializing the microcontroller** (*main.py*).
 This is **copied to the microcontroller** via Thonny and the microcontroller is then restarted.
 The IRIS data set and the DIABETES data set are available as example data sets.
 'ReLU', 'Leaky ReLU', 'Sigmoid', 'Softmax' and 'Tanh' are available as activation functions.
 If required, other classification data can be used for training in Python.
 For other data sets, the parameters of the neural networks must be adjusted accordingly.
-AI-ANNE-B(asic) works directly in Thonny and does not require a display or RGB matrix.
+AI-ANNE-B(asic) (*ai-anne-b.py*) works directly in Thonny and any Python environment, even without display, RGB matrix or dependencies.
 
 # License & Acknowledgment
 The current version 2.0 of AI-ANNE is available under **MIT License**.
@@ -51,13 +51,13 @@ In einem laufenden Forschungsvorhaben wird der Einsatz von KI-ENNA zur Detektion
 Im Ordner "python" befindet sich jeweils ein Datensatz im CSV-Format sowie Jupyter Notebooks für zwei Neuronale Netze zur Klassifikation.
 Die Datensätze wurden jeweils als Grundlage für das Training der Neuronalen Netze mit TensorFlow in Python verwendet.
 Die Parameter der Neuronalen Netze wurden daraufhin als TXT-Datei extrahiert.
-Im Ordner "micropython" befindet sich der **Code zur Initialisierung des Microcontrollers**.
+Im Ordner "micropython" befindet sich der **Code zur Initialisierung des Microcontrollers** (*main.py*).
 Dieser wird via Thonny **auf den Microcontroller kopiert** und dieser anschließend neu gestartet.
 Als Beispieldatensätze stehen der IRIS Datensatz und der DIABETES Datensatz zur Verfügung.
 'ReLU', 'Leaky ReLU', 'Sigmoid', 'Softmax' und 'Tanh' stehen als Aktivierungsfunktionen zur Verfügung.
 Bei Bedarf können andere Klassifikationsdaten für das Training in Python verwendet werden.
 Bei anderen Datensätzen sind die Parameter der Neuronalen Netze entsprechend anzupassen.
-KI-ENNA-B(asic) funktioniert direkt in Thonny und benötigt weder Display noch RGB-Matrix.
+KI-ENNA-B(asic) (*ai-anne-b.py*) funktioniert direkt in Thonny und jeder Python Umgebung, ganz ohne Display, RGB-Matrix oder Abhängigkeiten.
 
 # Lizenz & Danksagung
 In der aktuellen Version 2.0 steht KI-ENNA unter der **MIT Lizens** zur Verfügung.
