+++ ki-enna-b.py
@@ -1,5 +1,5 @@
 #################
-### AI-ANNE-B ###
+### KI-ENNA-B ###
 #################
 
 # Setup
