### 2025-04-20 #8039 @statistical-thinking
@kalpan80 Thanks for the quick review!

Since AI-ANNE can be put into operation simply by transferring main.py to the Raspberry Pi Pico and maps the exact results of a pre-trained neural network from Python to MicroPython via the transfer of weights and biases, I wouldn't know how to further demonstrate the **functionality** :-)

But I have added the transfer of main.py (with hardware) or ai-anne.py (without hardware) again on GitHub and added the above mentioned functionality in the section “How to Use and Support AI-ANNE”. I have also added the **Community Guidelines** there...

Thanks for your feedback!
