### 2025-04-20 #8039 @osorensen
Thanks @kalpan80. Do you have any specific suggestions or comments related to how @statistical-thinking can address the unchecked items on your review checklist?
