+++ README.md
@@ -81,4 +81,4 @@ Besonderer Dank gilt **@tonygod2** von instructables.com für seine hilfreichen 
 ------------------------------------------------------------------------------------------
 
 ### Application Example / Anwendungsbeispiel:
-![KI-ENNA](images/APPLICATION.png)
+![KI-ENNA](https://www.statistical-thinking.de/documents/ai-anne-demo.gif)
