+++ README.md
@@ -81,4 +81,4 @@ Besonderer Dank gilt **@tonygod2** von instructables.com für seine hilfreichen 
 ------------------------------------------------------------------------------------------
 
 ### Application Example / Anwendungsbeispiel:
-![KI-ENNA](https://www.statistical-thinking.de/documents/ai-anne-demo.gif)
+<img src="https://www.statistical-thinking.de/documents/ai-anne-demo.gif" alt="AI-ANNE_DEMO">
