### 2025-04-20 #8039 @kalpan80
Agree with @statistical-thinking on the functionality aspect of the library. Have marked it as completed.

Community guidelines - Any enthusiastic contributor can be allowed to create a PR for the KI.ENNA project. If the PR is found helpful, author can decide to merge it with the master branch.

Automated unit testing - We can leverage some of the pointers mentioned in below article.
https://admantium.medium.com/raspberry-pico-unit-test-framework-for-your-projects-f92623524446
