### 2025-04-21 #8039 @statistical-thinking
@kalpan80 @osorensen Thanks again for your input!

I have **added a test_ai_anne() function** in the basic code of AI-ANNE. This is **based on a small neural network** with just a **few neurons** and the **easy-to-understand sigmoid function**. The successful calculation of a Confusion Matrix and the Accuracy are defined as **test criteria**, as these are the goal of AI-ANNE.

I had considered implementing the Pi Pico's built-in LED as an additional signal in the test, but since not all users will have a Pi Pico and teaching with AI-ANNE can also take place only in Thonny as software environment, I decided against it :-)
