+++ README.md
@@ -81,4 +81,5 @@ Besonderer Dank gilt **@tonygod2** von instructables.com für seine hilfreichen 
 ------------------------------------------------------------------------------------------
 
 ### Application Example / Anwendungsbeispiel:
-![KI-ENNA](images/AI-ANNE-DEMO.gif)
+<img src= "images/AI-ANNE-DEMO.gif" alt="AI-ANNE-DEMI" style="border: 2px solid grey;">
+
