+++ README.md
@@ -80,5 +80,5 @@ Besonderer Dank gilt **@tonygod2** von instructables.com für seine hilfreichen 
 
 ------------------------------------------------------------------------------------------
 
-### Application Example / Anwendungsbeispiel:
-<kbd><img src= "images/AI-ANNE-DEMO.gif" alt="AI-ANNE-DEMI"><kbd>
+### Application Example / Anwendungsbeispiel
+![Avatar](images/AI-ANNE-DEMO.gif)
