### 2025-05-21 #8039 @kalpan80
@osorensen / @statistical-thinking - Please let me know if you have any other AI/ML papers to review. Would be happy to assist

Also, please let me know if anything need at my end to close out the review.
