### 2025-05-21 #8039 @statistical-thinking
@osorensen Is it necessary to change the title in this thread as well, as we did in the previous thread, to match the title of the paper?

@kalpan80 @samiralavi I am happy to announce that AI-ANNE has now won 3rd place in an inter-university award for innovative teaching, in addition to winning an award at my university of applied sciences. Next month AI-ANNE is going to the Helmholtz AI Conference in Germany. Exciting times... they seem to like the simple touch and feel as demonstrated in the GIF :-)
