### 2025-06-04 12b9291
Delete ai-anne-b.py
