### 2025-06-04 8191560
Delete ki-enna-b.py
