+++ ki-enna-b.py
@@ -1,0 +1,351 @@
+#################
+### KI-ENNA-B ###
+#################
+
+####################
+# (1) VORTRAINIERT #
+####################
+
+# Setup
+import math
+
+# ReLU
+def relu(x):
+    y = []
+    for i in range(len(x)):
+        if x[i] >= 0:
+            y.append(x[i])
+        else:
+            y.append(0)
+    return y
+
+# Leaky ReLU
+def leaky_relu(x, alpha=0.01):
+    p = []
+    for i in range(len(x)):
+        if x[i] >= 0:
+            p.append(x[i])
+        else:
+            p.append(alpha * x[i])
+    return p
+
+# Tanh
+def tanh(x):
+    t = [(math.exp(x[val]) - math.exp(-x[val])) / (math.exp(x[val]) + math.exp(-x[val])) for val in range(len(x))]
+    return t
+
+# Sigmoid
+def sigmoid(x):
+    z = [1 / (1 + math.exp(-x[val])) for val in range(len(x))]
+    return z
+
+# Softmax
+def softmax(x):
+    max_x = max(x[val])
+    exp_x = [math.exp(val - max_x) for val in range(len(x))]
+    sum_exp_x = sum(exp_x)
+    s = [j / sum_exp_x for j in exp_x]
+    return s
+
+# Single Neuron
+def neuron(x, w, b, activation):
+
+    tmp = zero_dim(x[0])
+
+    for i in range(len(x)):
+        tmp = add_dim(tmp, [(float(w[i]) * float(x[i][j])) for j in range(len(x[0]))])
+
+    if activation == "sigmoid":
+        yp = sigmoid([tmp[i] + b for i in range(len(tmp))])
+    elif activation == "relu":
+        yp = relu([tmp[i] + b for i in range(len(tmp))])
+    elif activation == "leaky_relu":
+        yp = relu([tmp[i] + b for i in range(len(tmp))])
+    elif activation == "tanh":
+        yp = tanh([tmp[i] + b for i in range(len(tmp))])
+    elif activation == "softmax":
+        yp = tanh([tmp[i] + b for i in range(len(tmp))])
+    else:
+        print("Function unknown!")
+
+    return yp
+
+# Mathematical Basics - I
+def zero_dim(x):
+    z = [0 for i in range(len(x))]
+    return z
+
+# Mathematical Basics - II
+def add_dim(x, y):
+    z = [x[i] + y[i] for i in range(len(x))]
+    return z
+
+# Mathematical Basics - III
+def zeros(rows, cols):
+    M = []
+    while len(M) < rows:
+        M.append([])
+        while len(M[-1]) < cols:
+            M[-1].append(0.0)
+    return M
+
+# Mathematical Basics - IV
+def transpose(M):
+    if not isinstance(M[0], list):
+        M = [M]
+    rows = len(M)
+    cols = len(M[0])
+    MT = zeros(cols, rows)
+    for i in range(rows):
+        for j in range(cols):
+            MT[j][i] = M[i][j]
+    return MT
+
+# Mathematical Basics - V
+def print_matrix(M, decimals=3):
+    for row in M:
+        print([round(x, decimals) + 0 for x in row])
+
+# Mathematical Basics - VI
+def dense(nunit, x, w, b, activation):
+    res = []
+    for i in range(nunit):
+        z = neuron(x, w[i], b[i], activation)
+        res.append(z)
+    return res
+
+# Include Parameters from TensorFlow
+w1 = [[-0.75323504, -0.25906014],
+      [-0.46379513, -0.5019245 ],
+      [ 2.1273055 ,  1.7724446 ],
[…]
