+++ paper/paper.md
@@ -18,7 +18,7 @@ bibliography: paper.bib
 
 # Summary
 
-Machine learning and deep learning are increasingly driving innovation across various fields [@LeCun2015-hn] and can also be transferred on microcontrollers [@Ray2022-oq], for example to process sensor data [@Cioffi2020-lq]. Since microcontrollers have limited computational resources [@Delnevo2023-ux], training or developing neural networks on microcontrollers remains a complex challenge [@Wulfert2024-qf]. However, MicroPython as programming language enables a resource-efficient implementation of neural networks on microcontrollers and provides insights into the fundamentals of neural networks in terms of explainable artificial intelligence [@Haque2023-du; @Meske2022-ke], while coding from scratch in MicroPython also allows for a streamlined and tailored approach [@Delnevo2023-ux]. AI-ANNE: (A) (N)eural (N)et for (E)xploration provides a framework that transfers neural networks from Python to MicroPython and enables the application of pre-trained TensorFlow and Keras models on microcontrollers as well as training directly on the microcontroler via forward and backward propagation. Furthermore, it enables users to explore the performance of neural networks while simultaneously the number of neurons and layers as well as the underlying activation functions can be adjusted easily in MicroPython, which makes it also suitable for didactic application [@Collier2024-ld; @Meske2022-ke; @Verma2022-ie; @Scherer2016-fy].
+Machine learning and deep learning are increasingly driving innovation across various fields [@LeCun2015-hn] and can also be transferred on microcontrollers [@Ray2022-oq], for example to process sensor data [@Cioffi2020-lq]. Since microcontrollers have limited computational resources [@Delnevo2023-ux], training or developing neural networks on microcontrollers remains a complex challenge [@Wulfert2024-qf]. However, MicroPython as programming language enables a resource-efficient implementation of neural networks on microcontrollers and provides insights into the fundamentals of neural networks in terms of explainable artificial intelligence [@Haque2023-du; @Meske2022-ke], while coding from scratch in MicroPython also allows for a streamlined and tailored approach [@Delnevo2023-ux]. AI-ANNE: (A) (N)eural (N)et for (E)xploration provides a framework that transfers neural networks from Python to MicroPython and enables the application of pre-trained TensorFlow and Keras models on microcontrollers as well as training directly on the microcontroller via forward and backward propagation. Furthermore, it enables users to explore the performance of neural networks while simultaneously the number of neurons and layers as well as the underlying activation functions can be adjusted easily in MicroPython, which makes it also suitable for didactic application [@Collier2024-ld; @Meske2022-ke; @Verma2022-ie; @Scherer2016-fy].
 
 # Statement of need
 
