### 2025-06-05 #8039 @statistical-thinking
Now forward and backward propagation is also possible directly on the microcontroller and I have stored the corresponding research poster from the Helmholtz AI Conference 2025 in the repository - yesterday at the conference was a huge success, despite the fact that EdgeAI or TinyML for embedded systems are still niche products :-)
