### 2025-06-06 #8039 @samiralavi
@statistical-thinking, congratulations on this work. I like how it is designed for educational purposes. I also would like to refer you to "Tensorflow lite for microcontrollers" (https://ai.google.dev/edge/litert/microcontrollers/overview), a subpart of the Tensorflow library that is designed to be for inferencing ANNs on low-end devices. You might get some ideas for the future. Your work meets the required criteria I was asked to review to be published in JOSS.
