+++ paper/paper.md
@@ -22,7 +22,7 @@ Machine learning and deep learning are increasingly driving innovation across va
 
 # Statement of need
 
-A previous investigation of skill requirements in artificial intelligence and machine learning job advertisements [@Verma2022-ie] served as a guide for the development of AI-ANNE: (A) (N)eural (N)et for (E)xploration. Therefore, it demystifies the mathematical principles underlying neural networks, allowing users to better understand the relationships between data, weights, biases and outputs [@Schmidt2020-ak; @Frank2020-rw; @Scherer2016-fy]. This includes reducing the number of layers, optimizing the activation functions, or using specialized operations to reduce computational load [@Sakr2021-mc]. Moreover, by developing the neural network directly in MicroPython, the model becomes highly portable and adaptable to various microcontroller architectures, such as the Raspberry Pi Pico or the Raspberry Pi Pico 2. Escpecially the step-by-step visibility in MicroPython helps users to identify and understand common challenges like vanishing or exploding gradients, while fostering a more intuitive grasp of how neural networks learn [@Kong2022-sk]. Debugging the code in MicroPython further reinforces this knowledge and provides insights into how errors propagate and are corrected during training. This is in accordance with the primary goals of explainable artificial intelligence [@Haque2023-du; @Schmidt2020-ak] and provides insights into artificial intelligence related responsibilities [@Collier2024-ld; @Frank2020-rw] as well as practical experiences [@Li2021-bn]. As such, implementing neural networks on microcontrollers with AI-ANNE: (A) (N)eural (N)et for (E)xploration shall promote necessary innovation and problem-solving skills [@Verma2022-ie].
+A previous investigation of skill requirements in artificial intelligence and machine learning job advertisements [@Verma2022-ie] served as a guide for the development of AI-ANNE: (A) (N)eural (N)et for (E)xploration. Therefore, it demystifies the mathematical principles underlying neural networks, allowing users to better understand the relationships between data, weights, biases and outputs [@Schmidt2020-ak; @Frank2020-rw; @Scherer2016-fy]. This includes reducing the number of layers, optimizing the activation functions, or using specialized operations to reduce computational load [@Sakr2021-mc]. Moreover, by developing the neural network directly in MicroPython, the model becomes highly portable and adaptable to various microcontroller architectures, such as the Raspberry Pi Pico or the Raspberry Pi Pico 2. Especially the step-by-step visibility in MicroPython helps users to identify and understand common challenges like vanishing or exploding gradients, while fostering a more intuitive grasp of how neural networks learn [@Kong2022-sk]. Debugging the code in MicroPython further reinforces this knowledge and provides insights into how errors propagate and are corrected during training. This is in accordance with the primary goals of explainable artificial intelligence [@Haque2023-du; @Schmidt2020-ak] and provides insights into artificial intelligence related responsibilities [@Collier2024-ld; @Frank2020-rw] as well as practical experiences [@Li2021-bn]. As such, implementing neural networks on microcontrollers with AI-ANNE: (A) (N)eural (N)et for (E)xploration shall promote necessary innovation and problem-solving skills [@Verma2022-ie].
 
 # Basic elements
 AI-ANNE: (A) (N)eural (N)et for (E)xploration includes the activation functions ReLU, Leaky ReLU, Tanh, Sigmoid and Softmax as pre-installed activation functions. Additional activation functions can be flexibly added in MicroPython. Furthermore, neurons, matrices and matrix transposition as well as the density of neural networks are already pre-programmed in MicroPython. A pre-installed confusion matrix makes it possible to evaluate the performance of different neural networks.
