### 2025-06-07 #8039 @osorensen
@statistical-thinking, at this point could you:
